move report event to ray utils
